修复click_img,guochang 把结束线程逻辑移动至_get_screen中。
